Split AzurePlugin into java and rider parts
