+++ PluginsAndFeatures/azure-toolkit-for-intellij/rider/resources/META-INF/plugin.xml
@@ -24,7 +24,13 @@
     </component>
   </application-components>
 
-  <project-components/>
+  <project-components>
+    <component>
+      <implementation-class>com.microsoft.intellij.AzureRiderPlugin</implementation-class>
+      <loadForDefaultProject/>
+    </component>
+  </project-components>
+
   <actions/>
 
 </idea-plugin>
